Move java files into packages
