+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".FeedActivity"
+            android:name=".activity.FeedActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
@@ -25,20 +25,20 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".DetailActivity"
+            android:name=".activity.DetailActivity"
             android:label="@string/title_activity_detail"
             android:theme="@style/AppTheme.NoActionBar"
-            android:parentActivityName="uk.ac.kcl.stranders.hitour.FeedActivity">
+            android:parentActivityName=".activity.FeedActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="uk.ac.kcl.stranders.hitour.FeedActivity" />
+                android:value=".activity.FeedActivity" />
         </activity>
         <activity
-            android:name=".ScanningActivity"
-            android:parentActivityName="uk.ac.kcl.stranders.hitour.FeedActivity">
+            android:name=".activity.ScanningActivity"
+            android:parentActivityName=".activity.FeedActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="uk.ac.kcl.stranders.hitour.FeedActivity" />
+                android:value=".activity.FeedActivity" />
         </activity>
     </application>
