+++ app/src/main/res/layout/activity_feed.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="uk.ac.kcl.stranders.hitour.FeedActivity">
+    tools:context=".activity.FeedActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
