+++ app/src/main/res/layout/activity_scanning.xml
@@ -12,7 +12,7 @@
         android:paddingLeft="@dimen/activity_horizontal_margin"
         android:paddingRight="@dimen/activity_horizontal_margin"
         android:paddingTop="@dimen/activity_vertical_margin"
-        tools:context="uk.ac.kcl.stranders.hitour.ScanningActivity">
+        tools:context=".activity.ScanningActivity">
 
         <com.journeyapps.barcodescanner.CompoundBarcodeView
             android:id="@+id/zxing_barcode_scanner"
