+++ app/src/main/res/layout/content_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="uk.ac.kcl.stranders.hitour.DetailActivity"
+    tools:context=".activity.DetailActivity"
     tools:showIn="@layout/activity_detail">
 
     <LinearLayout
