+++ app/src/main/res/layout/content_feed.xml
@@ -7,7 +7,7 @@
     android:paddingTop="8dp"
     android:background="#ECEFF1"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="uk.ac.kcl.stranders.hitour.FeedActivity"
+    tools:context=".activity.FeedActivity"
     tools:showIn="@layout/activity_feed">
 
     <android.support.v7.widget.RecyclerView
